new kernels, modify constructor vli_gpu_vector

+++ large_integer/VLI/detail/vli_size_param.hpp
@@ -34,7 +34,7 @@
 /**
 * type of the vli
 */
-typedef unsigned int type;
+typedef unsigned long int type;
 
 /** 
 *END FIX YOUR PARAMETERS
